requirements.txt
discord.py
PyYAML==6.0.3
requests==2.32.5
yt-dlp
pynacl